tty: dup /dev/null over stdio for (detach && !terminal)
This cleans up some of the terminal=false semantics for detached
containers. One of the biggest issues with running `runc create` inside
a terminal is that it gave the container process access to a host tty.
While this is not a security issue in of itself, it is quite concerning
(and similar bugs have caused security issues in LXC in the past). Not
to mention that this meant that the user's shell was borked -- something
which should be fixed anyway.

Signed-off-by: Aleksa Sarai <asarai@suse.de>