README: Document the alternate-week meeting times
On Thu, Mar 09, 2017 at 08:37:55AM -0700, Chris Aniszczyk wrote [1]:
> summarizing the discussion, how about we just alternative the time like
> Jonathan discussed?
>
> I believe that's the only fair thing to do and it's reassuring to hear from
> people like Phil who would be able to make the time along with others.
>
> I'm fine with the current time (1000AEST/1500PST/0000CET) and then
> 0400AEST/0900PST/1800CET

The 8am Pacific time ended up working out better than 9am for Jonathan
and Samuel [2,3], so I've used that instead of 9am.  The 8am slot was
confirmed as the most popular slot in a Doodle poll [4], with the
following folks approving that slot:

* David Lyle
* George Lestaris
* Jonathan Boulle
* Julz
* Michael Crosby
* Mike Brown
* Mrunal Patel
* Phil Estes
* Rob Dolin
* Samuel Ortiz
* Stephen Day
* Stephen Walli
* Vincent Batts
* W. Trevor King

Removing those folks, the second most popular slot is 5pm Pacific,
with the following folks approving that slot:

* Aleksa Sarai
* Keyang Xie
* Lei Jitang
* Ma Shimiao
* Qiang Huang

Stephen and Mrunal approved both slots, and since they frequently
anchor the runtime and image conversations respectively, there should
be sufficient continuity between the two meetings.

The only person voting in the Doodle poll who didn't approve either
slot is Tianon.

Folks with a POSIX ‘date’ command can find the week number with [5]:

  $ date +%V

There may be some doubling up around the end of the year, but we're
usually canceling meetings around then anyway.

The 8am Pacific meeting gets the odd slot because it's Europe-friendly
and lots of folks will be in Europe on 2017-03-29 for KubeCon [6].

I'd be happier with meeting times anchored to UTC to make life easier
for folks outside of the US, but one change at a time ;).

Future bumps to meeting.ics should bump LAST-MODIFIED [7] or DTSTAMP
[8] for any altered components.  We can't use DTSTAMP in the VEVENT
because VEVENTs require DTSTAMP [9].

The timezone entry is based on the America/New_York example from [10].

Figuring out a single RRULE to cover both meeting times was beyond my
abilities, and while RFC 2445 allowed multiple RRULEs in a single
VEVENT [11,12], RFC 5545 does not [13].  Something like:

  RRULE:FREQ=YEARLY;BYDAY=WE;BYHOUR=8,17;BYSETPOS=1,4,5,8,9,...

should be legal (at least for 2017), but Google Calendar [14] doesn't
seem to respect BYHOUR expansion, and ICAL.js [15] doesn't seem to
respect the BYSETPOS limit, so I gave it up and went with two events.

To stick strictly to the ISO weeks we could use:

  RRULE:FREQ=YEARLY;BYDAY=WE;BYWEEKNO=13,15,17,19,21,23,25,27,29,31,33,35,
   37,39,41,43,45,47,49,51

and:

  RRULE:FREQ=YEARLY;BYDAY=WE;BYWEEKNO=14,16,18,20,22,24,26,28,30,32,34,36,
   38,40,42,44,46,48,50,52

but that's tedious to type, and folks probably don't care all that
much about ISO weeks.  I've gone with WEEKLY and INTERVAL=2 to give us
something that might survive the end of the year.

The ICS was validated with [16].

The CRLF line endings are intentional [17], and the .gitattributes
entry ensures we keep them.  The committed files will still have LF
endings, which can confuse 'git diff ...', but you can use
--ignore-space-at-eol to see what really changed.

[1]: https://groups.google.com/a/opencontainers.org/d/msg/dev/p0mTOspVgd0/mh7FYse2BAAJ
     Subject: Re: Moving the OCI Call (again)
     Date: Thu, 9 Mar 2017 08:37:55 -0700
     Message-ID: <CAJg1wMTCGEFRuKoKBEbUPdho82TVH8sPZdGORK_NA2vCNe+w9w@mail.gmail.com>
[2]: https://groups.google.com/a/opencontainers.org/d/msg/dev/p0mTOspVgd0/ULXnARy9BAAJ
     Subject: Re: Moving the OCI Call (again)
     Date: Thu, 9 Mar 2017 18:33:34 +0100
     Message-ID: <CAPWU_0rByhFp=jQQ6cvagHJuYmeTvN7T1zAW+oZR3=F1W8b_rw@mail.gmail.com>
[3]: https://github.com/opencontainers/runtime-spec/pull/719#pullrequestreview-26109314
[4]: http://doodle.com/poll/zu664785gb59pwkg
[5]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/date.html
[6]: http://ircbot.wl.linuxfoundation.org/meetings/opencontainers/2017/opencontainers.2017-03-22-21.00.log.html
[7]: https://tools.ietf.org/html/rfc5545#section-3.8.7.3
[8]: https://tools.ietf.org/html/rfc5545#section-3.8.7.2
[9]: https://tools.ietf.org/html/rfc5545#section-3.6.1
[10]: https://tools.ietf.org/html/rfc5545#page-69
[11]: https://tools.ietf.org/html/rfc2445#section-4.6.1
[12]: https://tools.ietf.org/html/rfc2445#section-4.8.5.4
[13]: https://tools.ietf.org/html/rfc5545#appendix-A.1
[14]: https://calendar.google.com/
[15]: http://mozilla-comm.github.io/ical.js/
[16]: https://icalendar.org/validator.html
[17]: https://tools.ietf.org/html/rfc5545#section-3.1

Signed-off-by: W. Trevor King <wking@tremily.us>