runtime: Drop "Barring access control concerns"
This wording landed without comment as part of 7117ede7 (Expand on the
definition of our ops, 2015-10-13, #225).  However, I'm not entirely
clear on the exception it's making.  It may be trying to say something
like:

  Just because you were authorized to manage that container when you
  created it doesn't mean you're still authorized to perform operation
  X on it now.  Maybe you've lost privileges in the meantime.

But as far as compliance testing is concerned, the same test harness
will be calling 'create' and the subsequent operations.  That harness
will be reporting MUST violations if the runtime refuses a subsequent
operation, and removing the access-control loophole makes it more
obvious that the runtime's refusal is non-compliant.

Signed-off-by: W. Trevor King <wking@tremily.us>