config: Clarify mounts[].source relative path anchor
Borrowing language from root.path for absolute/relative.

I think we also need to talk about 'mounts' being in the container
mount namespace (after creating it with the runtime mount namespace as
a seed) and before the pivot into root.path.  But Michael and/or
Mrunal are going to file follow-up work to address that.

Signed-off-by: W. Trevor King <wking@tremily.us>