runtime-linux: Condition /proc/self/fd symlinks on source existence
Since 279c3c09 (linux: relax filesystem requirements for container,
2017-01-23, #666) it's no longer guaranteed that /proc will exist.
And there doesn't seem to be much point in requiring symlinks which
will be known broken.

This commit also tightens the timing.  Before it was just "after the
container has `/proc` mounted", which could have happened during the
'delete' operation (if the container authors wanted to be especially
ornery).  With this commit, I've put the creation in step 2 of the
lifecycle.  And within step 2, it happens after 'mounts' has been
processed.

Signed-off-by: W. Trevor King <wking@tremily.us>