meeting: Add UberConference information
And bump DTSTAMP for the touched VEVENTS.

This is less DRY, but makes life easier for folks pulling up the
calendar a minute before the meeting, since they don't have to click
through to the README section.  Also requested by Vincent [1] ;).

The LOCATION entry [2] seems like a reasonable location for the
conference-call page, so I've used that.  Note that the URI there is a
text value; LOCATION supports URIs in ALTREP [2], but the
UberConference page isn't an "alternate text representation for the
property value" [3].  The URI in DESCRIPTION is a fallback for clients
that don't support/display LOCATION (I'm not sure if any exist, but
it's easy to play it safe).  I couldn't find a more structured
location for the phone number, unless we also wanted to put that in
LOCATION.

The ICS was validated with [4].

[1]: http://ircbot.wl.linuxfoundation.org/eavesdrop/%23opencontainers/%23opencontainers.2017-04-05.log.html#t2017-04-05T17:55:25
[2]: https://tools.ietf.org/html/rfc5545#section-3.8.1.7
[3]: https://tools.ietf.org/html/rfc5545#section-3.2.1
[4]: https://icalendar.org/validator.html

Signed-off-by: W. Trevor King <wking@tremily.us>