meeting: Include README link in DESCRIPTION body
And bump DTSTAMP for the touched VEVENTS.

com.android.calendar version 7.1.1 only displays the DESCRIPTION body,
and does not provide a link to ALTREP (in fact, I don't see any
instances of ALTREP in the source [1]).  Including the README link in
the DESCRIPTION body gives folks using that calendar application an
easy way to get to the README section.

The ICS was validated with [2].

[1]: https://android.googlesource.com/platform/packages/apps/Calendar/+/android-7.1.1_r38
[2]: https://icalendar.org/validator.html

Signed-off-by: W. Trevor King <wking@tremily.us>