validator.go: check the given net namespace is host net namespace
If the network namespace is not a symlink, it still could
be a host network namespace. For exampe, in docker, the
host network namespace path is `/var/run/docker/netns/default`,
and it's a mount point of host network namespace.

```
$ mount | grep default
nsfs on /run/docker/netns/default type nsfs (rw)
```
This will make the docker can set `sysctl` when use
host network namespace and this will cause security
issue if we update the runc of docker.

Signed-off-by: Lei Jitang <leijitang@huawei.com>