config-linux: RFC 2119 wording for intelRdt
So we can compliance-test runtimes for these settings.

Also remove the tutorial, since the kernel docs should provide
sufficient documentation on that front.  The kernel can be patched if
they do not, and we do not include tutorials for other config-linux
settings in this spec.

The updated example was recommended by Xiaochen to compensate for the
removed inline tutorial [1].

[1]: https://github.com/opencontainers/runtime-spec/pull/787#discussion_r114254422

Signed-off-by: W. Trevor King <wking@tremily.us>