config-linux: Use the implicit link name shortcut
727e64d9 (config-linux.md: fix wrong link description about cgroups,
2017-05-14, #826) added a redundant link name for libseccomp.  This
commit restores the old implicit link name shortcut [1] there and
updates the other entries (all of which were for the filesystems table
in config-linux) that could also use that shortcut.  This is obviously
not a major change, but I think it makes the source easier to read by
minimizing the markup so you can focus on the text.

[1]: https://daringfireball.net/projects/markdown/syntax#link

Signed-off-by: W. Trevor King <wking@tremily.us>