bundle: Move 'rootfs' recommendation to config.md
Anyone setting 'rootfs.path' should be aware of this advice,
regardless of whether they're the ones composing the bundle or not.

Use the RFC 2119 "SHOULD" (vs. the old lowercase "should") for this
recommendation.  The SHOULD semantics make sense and using SHOULD
avoids confusing readers ("did they mean to SHOULD this?").

Also drop the "While the name of this directory may be arbitrary"
caveat, because SHOULD already implies "but you can pick another
directory name if you want".

Note that the "MUST be present in a single directory" line which is
still in bundle.md forbids you from picking "foo/bar", "../foo", or
other paths that do not point at a direct child of the bundle
directory.  I don't like that direct-child restriction [1], but I'm
not touching it in this commit.

[1]: https://github.com/opencontainers/runtime-spec/pull/469

Signed-off-by: W. Trevor King <wking@tremily.us>