config: Require hook support on Linux and Solaris
Clarifying the language from 28e8f68 (Hooks to linux,solaris and doc
clarification, 2017-05-24, #855), which did not say which platforms
support hooks and which don't.  Without something like this commit,
there's no clear way for config authors to know if their runtime will
support hooks or not, and there was previous agreement that that sort
of ambiguity was not helpful [1].  This also gives normative Markdown
grounding for the Go platform tag added in 28e8f68.

[1]: https://github.com/opencontainers/runtime-spec/pull/472#r78088802

Signed-off-by: W. Trevor King <wking@tremily.us>