config: Use "POSIX" instead of "Linux and Solaris"
As requested by Michael [1], this reduces the number of changes
required if BSD or some such is added to the config.

This change leaves a bit of a gap between the platforms listed in
spec.md, since readers have to figure out on their own that the POSIX
properties apply to the linux and solaris platforms (and potentially
other future platforms).  But Michael felt like it's ok to leave that
gap, at least for now [2].

I've used "On POSIX platforms" and similar (instead of "On POSIX")
because, as Tianon points out [3], there is some space between
POSIX-the-spec and platforms which implement it.

[1]: https://github.com/opencontainers/runtime-spec/pull/838#discussion_r119729323
[2]: http://ircbot.wl.linuxfoundation.org/meetings/opencontainers/2017/opencontainers.2017-06-01-17.41.log.html#l-60
[3]: https://github.com/opencontainers/runtime-spec/pull/838#discussion_r123279055

Signed-off-by: W. Trevor King <wking@tremily.us>