Fix non-Unix build of libcontainer/user
Commit 3d7cb4293c15 ("Move libcontainer to x/sys/unix") switched from
`syscall.Get[ug]id` to `unix.Get[ug]id` where `unix.*` is (naturally) not
available on non-Unix platforms such as Windows.

Move the offending code into a Unix specific file and provide stubs in the
_unsupported.go case.

The stubs use `-1` for UID and GID which is what syscall_windows.go was
returning prior to 3d7cb4293c15, and these are laundered through `Lookup[GU]id`
to provide the same behaviour as before.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>