config: Simplify title to "Configuration"
The capitalization in "Container Configuration file" (which we've used
since 70372d38, *.md: update TOC and links, 2015-09-10, #176) was
halfway between "Title Case" and "Sentence case".  The current spec
isn't particularly consistent (e.g. we have both "Specification
version" and "POSIX-platform Mounts"), but the ToC has used
"Configuration" for this file since e7be40f0 (Cleanup the spec a bit
to remove WG/git text that's not really part of the spec, 2016-11-14,
#626) so dodge the sentence/title issue and use that here too.

Signed-off-by: W. Trevor King <wking@tremily.us>