Fix cgroup: add check task exists after writing freezer.state
Description:
  Example: we start a container which just run `sleep 3` command in
container. After 3 seconds, if we try to pause it (container exit
process in shim/container is not finished), the cgroup file exists.

So we could update the cgroup file successfully, but the container
process `sleep 3` has already exited.

Then we got an 'Exited' container with paused state in docker.

So I think we should check if the `tasks` is empty or not after update
`freezer.state`.

Signed-off-by: Wentao Zhang <zhangwentao234@huawei.com>