rootfs: preserve old mount flags when remounting bindmount
Fixes the case where a bind-mount is being defined in a user namespaced
container which requires a remount. Previously this code would attempt
to do a MS_REMOUNT that potentially dropped mount flags that were
inherited through the MS_BIND. Resolve this by intentionally adding the
mount flags to the mount configuration in that scenario (if the remount
initially f ailed).

This restriction comes from CL_UNPRIVILEGED being set when a mount
namespace is created in a user namespace different from the source mount
namespace's owner[1]. However, this restriction only locks a certain set
of flags. We only inherit those flags if they are set.

[1]: http://elixir.free-electrons.com/linux/v4.13.4/source/fs/namespace.c#L1033

Signed-off-by: Aleksa Sarai <asarai@suse.de>